+++ miniprogram/pages/order-result/index.js
@@ -2,6 +2,7 @@ Page({
   data: {
     orderNo: '',
     orderScene: '',
+    orderSceneLabel: '',
     payableAmount: '',
     contactPhone: '',
     roomDeliveryNotice: ''
@@ -11,6 +12,7 @@ Page({
     this.setData({
       orderNo: options.orderNo || '',
       orderScene: options.orderScene || '',
+      orderSceneLabel: this.formatOrderScene(options.orderScene),
       payableAmount: options.payableAmount || '0.00'
     })
     this.loadConfig()
@@ -35,6 +37,18 @@ Page({
     })
   },
 
+  goRoomDining() {
+    wx.navigateTo({
+      url: '/pages/room/index'
+    })
+  },
+
+  goServiceCenter() {
+    wx.switchTab({
+      url: '/pages/mine/index'
+    })
+  },
+
   goOrders() {
     wx.navigateTo({
       url: '/pages/order/list/index'
@@ -52,5 +66,14 @@ Page({
     wx.makePhoneCall({
       phoneNumber: this.data.contactPhone
     })
+  },
+
+  formatOrderScene(value) {
+    const map = {
+      NORMAL: '普通点餐',
+      ROOM_DELIVERY: '客房送餐',
+      PRIVATE_ROOM_PREORDER: '包间预点菜'
+    }
+    return map[value] || value || '未知场景'
   }
 })